Move exceptions to separate directory
