+++ src/ActiveMQ.Net/AutoRecovering/AutoRecoveringProducer.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Threading;
 using System.Threading.Tasks;
+using ActiveMQ.Net.Exceptions;
 using Microsoft.Extensions.Logging;
 using Nito.AsyncEx;
