+++ src/ActiveMQ.Net/Builders/ProducerBuilder.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Threading;
 using System.Threading.Tasks;
+using ActiveMQ.Net.Exceptions;
 using Amqp;
 using Amqp.Framing;
 using Microsoft.Extensions.Logging;
