+++ src/ActiveMQ.Net/Exceptions/CreateConnectionException.cs
@@ -1,0 +1,7 @@
+﻿namespace ActiveMQ.Net.Exceptions
+{
+    public class CreateConnectionException
+    {
+        
+    }
+}
