+++ src/ActiveMQ.Net/Exceptions/MessageSendException.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using Amqp.Framing;
 
-namespace ActiveMQ.Net
+namespace ActiveMQ.Net.Exceptions
 {
     public class MessageSendException : Exception
     {
